--- src/main/java/persistence/sql/ddl/DataType.java
@@ -1,24 +1,0 @@
-package persistence.sql.ddl;
-
-public enum DataType {
-    String,
-    Long,
-    Integer,
-    ;
-
-    DataType() {
-    }
-
-    public static String getDBType(String dataType) {
-        if (dataType.equals(DataType.String.name())) {
-            return "varchar(255)";
-        }
-        if (dataType.equals(DataType.Long.name())) {
-            return "bigint";
-        }
-        if (dataType.equals(DataType.Integer.name())) {
-            return "int";
-        }
-        throw new IllegalArgumentException("Not Supported DataType");
-    }
-}
